refactor: handle file packages and delete un used files and code.
