+++ ui/src/main/java/com/amsterdam/ui/components/DialogHeaderSection.kt
@@ -6,7 +6,8 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
-import com.amsterdam.designsystem.components.IconButton
+import com.amsterdam.designsystem.R
+import com.amsterdam.designsystem.components.buttons.IconButton
 import com.amsterdam.designsystem.components.Text
 import com.amsterdam.designsystem.theme.AppTheme
 
@@ -26,7 +27,7 @@ fun DialogHeaderSection(
         )
 
         IconButton(
-            painter = painterResource(com.amsterdam.designsystem.R.drawable.ic_cancel),
+            painter = painterResource(R.drawable.ic_cancel),
             contentDescription = null,
             tint = AppTheme.color.title,
             onClick = onDismiss,
