+++ ui/src/main/java/com/amsterdam/ui/components/selection/AnswerSelectionItem.kt
@@ -18,8 +18,8 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import com.amsterdam.designsystem.R
 import com.amsterdam.designsystem.components.Icon
-import com.amsterdam.designsystem.components.RadioButton
-import com.amsterdam.designsystem.components.RadioState
+import com.amsterdam.designsystem.components.buttons.RadioButton
+import com.amsterdam.designsystem.components.buttons.RadioState
 import com.amsterdam.designsystem.components.Text
 import com.amsterdam.designsystem.theme.AflamiTheme
 import com.amsterdam.designsystem.theme.AppTheme
